Pull out API types
